agent/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-cartesia
python-dotenv
requests
fastapi 
uvicorn 
passlib 
python-dotenv
"pymongo[srv]"==3.12

